+++ src/main/java/com/besson/arknights/block/custom/pizzeria/PizzeriaCupboard.java
@@ -1,0 +1,34 @@
+package com.besson.arknights.block.custom.pizzeria;
+
+import com.besson.arknights.block.FurnitureHorizontalFacingBlock;
+import com.besson.arknights.block.ModAbstractChestBlock;
+import net.minecraft.block.Block;
+import net.minecraft.block.BlockState;
+import net.minecraft.block.ShapeContext;
+import net.minecraft.block.entity.BlockEntity;
+import net.minecraft.block.entity.BlockEntityType;
+import net.minecraft.util.math.BlockPos;
+import net.minecraft.util.shape.VoxelShape;
+import net.minecraft.world.BlockView;
+import org.jetbrains.annotations.Nullable;
+
+import java.util.function.Supplier;
+
+public class PizzeriaCupboard extends ModAbstractChestBlock<PizzeriaCupboardBE> {
+
+    private static final VoxelShape SHAPE = Block.createCuboidShape(0, 0, 0, 16,16,16);
+
+    public PizzeriaCupboard(Settings settings, Supplier<BlockEntityType<? extends PizzeriaCupboardBE>> blockEntityTypeSupplier) {
+        super(settings, blockEntityTypeSupplier);
+    }
+
+    @Override
+    public VoxelShape getOutlineShape(BlockState state, BlockView world, BlockPos pos, ShapeContext context) {
+        return SHAPE;
+    }
+
+    @Override
+    public @Nullable BlockEntity createBlockEntity(BlockPos pos, BlockState state) {
+        return new PizzeriaCupboardBE(pos, state);
+    }
+}